requirements.txt
asgiref==3.6.0
autopep8==1.5.2
Django==4.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.2.2
pycodestyle==2.6.0
PyJWT==2.6.0
pytz==2020.1
sqlparse==0.3.1
